refactor: 성경통독 ui 수정
